requirements.txt
# Phase 1: Core API + LLM + Vector DB
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
langchain==0.1.7
langchain-openai==0.0.5
qdrant-client==1.7.3
openai==1.12.0
python-dotenv==1.0.1
sentence-transformers==2.5.1

# Phase 2: Streaming
confluent-kafka==2.3.0

# Phase 3: RL
gymnasium==0.29.1
ray[rllib]==2.9.3

# Phase 4: Observability
prometheus-client==0.20.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
